Fix an old inconsistency in the native 'L4D2Direct_GetSpawnTimer'.
This commit will force people to update their plugin code, as using this native feature incorrectly can break game data and cause bugs. It's been clear for a long time that it's not a countdowntimer, but a simple float, but if you don't update the code, the problems will persist.